tweaks

+++ src/core/cgl_shader.js
@@ -52,7 +52,7 @@ CGL.Shader = function(_cgl, _name) {
     this._inverseViewMatrixUniform = null;
 
     this._attrVertexPos = -1;
-    this.precision = _cgl.patch.config.glslPrecision||'mediump';
+    this.precision = _cgl.patch.config.glslPrecision||'highp';
 
     this._pMatrixState =-1;
     this._vMatrixState =-1;
